% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymmetric_sync.R
\name{common_files_asym_sync_to_right}
\alias{common_files_asym_sync_to_right}
\title{Partial asymmetric synchronization to right (update common files)}
\usage{
common_files_asym_sync_to_right(
  left_path = NULL,
  right_path = NULL,
  sync_status = NULL,
  by_date = TRUE,
  by_content = FALSE,
  recurse = TRUE,
  force = TRUE,
  backup = FALSE,
  backup_dir = "temp_dir",
  verbose = getOption("syncdr.verbose")
)
}
\arguments{
\item{left_path}{Path to the left/first directory.}

\item{right_path}{Path to the right/second directory.}

\item{sync_status}{Object of class "syncdr_status", output of \code{compare_directories()}.}

\item{by_date}{logical, TRUE by default}

\item{by_content}{logical, FALSE by default}

\item{recurse}{logical, TRUE by default.
If recurse is TRUE: when copying a file from source folder to destination folder, the file will be copied into the corresponding (sub)directory.
If the sub(directory) where the file is located does not exist in destination folder (or you are not sure), set recurse to FALSE,
and the file will be copied at the top level}

\item{force}{Logical. If TRUE (by default), directly perform synchronization of the directories.
If FALSE, Displays a preview of actions and prompts the user for confirmation before proceeding. Synchronization is aborted if the user does not agree.}

\item{backup}{Logical. If TRUE, creates a backup of the right directory before synchronization. The backup is stored in the location specified by \code{backup_dir}.}

\item{backup_dir}{Path to the directory where the backup of the original right directory will be stored. If not specified, the backup is stored in temporary directory (\code{tempdir}).}

\item{verbose}{logical. If TRUE, display directory tree before and after synchronization. Default is FALSE}
}
\value{
Invisible TRUE indicating successful synchronization.
}
\description{
Partially synchronize right directory based on left one -i.e., the function will:
\itemize{
\item for common_files:
\itemize{
\item if by date only: copy files that are newer in left to right
\item if by date and content: copy files that are newer and different in left to right
\item if by content only: copy files that are different in left to right
}
\item for non common files, nothing changes: i.e.,
\itemize{
\item disregard those files that are only in left
\item keep in right those files that are only in right (i.e., files 'missing in left')
}
}
}
\examples{
# Asymmetric synchronization of common files

\donttest{
e <- toy_dirs()
left  <- e$left
right <- e$right
# Synchronize common files by content only
common_files_asym_sync_to_right(
  left_path  = left,
  right_path = right,
  by_date    = FALSE,
  by_content = TRUE
)
}
}
