% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action_functions.R
\name{copy_files_to_left}
\alias{copy_files_to_left}
\title{Copy files from right directory to left directory}
\usage{
copy_files_to_left(
  left_dir,
  right_dir,
  files_to_copy,
  recurse = TRUE,
  overwrite = TRUE
)
}
\arguments{
\item{left_dir}{Path of the left (destination) directory.}

\item{right_dir}{Path of the right (source) directory.}

\item{files_to_copy}{Data frame containing paths of files to copy and their synchronization status.}

\item{recurse}{Logical, default is TRUE.
\itemize{
\item If TRUE: Files are copied into corresponding subdirectories in the left directory. If a subdirectory doesn't exist in the left directory, it will be created.
\item If FALSE: Files are copied to the top level of the left directory.
}}

\item{overwrite}{Logical, default is TRUE. If TRUE, existing files at the
destination are overwritten. If FALSE, existing destination files are
preserved and the copy is skipped.}
}
\value{
Invisible TRUE. Aborts with an error if the destination directory
is not writable (VUL-27). If individual file copies fail, a warning is
issued listing all failed files and copying continues for the remainder
(VUL-24/VUL-26).
}
\description{
This function copies files from a right (source) directory to a left (destination) directory based on a provided data frame containing file paths and synchronization status.
}
\details{
The function performs the following steps:
\enumerate{
\item Checks if the source (right) and destination (left) directories exist and creates the destination directory if it doesn't already exist.
\item Copies files from the right directory to the corresponding subdirectory in the left directory based on the provided file paths.
}
}
\keyword{internal}
