% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_dirs.R
\name{copy_temp_environment}
\alias{copy_temp_environment}
\title{Create a temporary copy of .syncdrenv to test functions}
\usage{
copy_temp_environment()
}
\value{
A list of temporary paths \code{left} and \code{right}.
}
\description{
This function creates a copy of the original environment, allowing tests to be executed without modifying the original environment.
}
