% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{directory_info}
\alias{directory_info}
\title{Retrieve information about files in a directory}
\usage{
directory_info(dir, recurse = TRUE)
}
\arguments{
\item{dir}{A character string representing the path of the directory.}

\item{recurse}{Logical. If TRUE, fully recurse into subdirectories. If a positive number,
specifies the number of levels to recurse.}
}
\value{
A data frame containing detailed information about the files in the directory,
including all information produced by \code{fs::file_info()}.
}
\description{
This function retrieves information about files in a specified directory
}
\keyword{internal}
