% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_functions.R
\name{display_dir_tree}
\alias{display_dir_tree}
\title{Display tree structure of one (or two) directory}
\usage{
display_dir_tree(path_left = NULL, path_right = NULL, recurse = TRUE)
}
\arguments{
\item{path_left}{path of left directory}

\item{path_right}{path of right directory}

\item{recurse}{logical, default to TRUE: show also sub-directories}
}
\value{
directories tree
}
\description{
Display tree structure of one (or two) directory
}
\examples{
# Create a temporary directory structure

\donttest{
e <- toy_dirs()
left  <- e$left
right <- e$right

display_dir_tree(
  path_left  = left,
  path_right = right
)

display_dir_tree(path_right = right)
}
}
