% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_functions.R
\name{display_file_actions}
\alias{display_file_actions}
\title{Display file actions in table}
\usage{
display_file_actions(path_to_files, directory, action = c("copy", "delete"))
}
\arguments{
\item{path_to_files}{Data frame containing the paths to the files. The data frame should have a column named "Paths".}

\item{directory}{Character string specifying path to directory where action is taken}

\item{action}{Character vector specifying the action to be performed on the files.
Options are "copy" (default) or "delete".}
}
\value{
console-friendly table with files and actions
}
\description{
This function displays actions (either "copy" or "delete") to be performed on a list of files.
}
\keyword{internal}
