% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{hash_files}
\alias{hash_files}
\title{Hash files by content}
\usage{
hash_files(files_path, verbose = getOption("syncdr.verbose"))
}
\arguments{
\item{files_path}{character vector of paths of files to hash}

\item{verbose}{logical; if TRUE display progress status of hashing. Default is FALSE}
}
\value{
hashes of files
}
\description{
Hash files by content
}
\keyword{internal}
