% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{hash_files_contents}
\alias{hash_files_contents}
\title{Hash content of files of two directories under comparison, say left and right}
\usage{
hash_files_contents(left_path, right_path)
}
\arguments{
\item{left_path}{path of files in left directory}

\item{right_path}{path of files in right directory}
}
\value{
list of hashes of files from left paths and hashes of files from right paths
}
\description{
Hash content of files of two directories under comparison, say left and right
}
\keyword{internal}
