% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymmetric_sync.R
\name{partial_update_missing_files_asym_to_right}
\alias{partial_update_missing_files_asym_to_right}
\title{Partial asymmetric asymmetric synchronization of non common files}
\usage{
partial_update_missing_files_asym_to_right(
  left_path = NULL,
  right_path = NULL,
  sync_status = NULL,
  recurse = TRUE,
  force = FALSE,
  backup = FALSE,
  backup_dir = "temp_dir",
  overwrite = TRUE,
  verbose = getOption("syncdr.verbose")
)
}
\arguments{
\item{left_path}{Path to the left/first directory.}

\item{right_path}{Path to the right/second directory.}

\item{sync_status}{Object of class "syncdr_status", output of \code{compare_directories()}.}

\item{recurse}{logical, TRUE by default.
If recurse is TRUE: when copying a file from source folder to destination folder, the file will be copied into the corresponding (sub)directory.
If the sub(directory) where the file is located does not exist in destination folder (or you are not sure), set recurse to FALSE,
and the file will be copied at the top level}

\item{force}{Logical. If FALSE (default), displays a preview of actions and prompts the user for confirmation before proceeding. Synchronization is aborted if the user does not agree. If TRUE, directly performs synchronization without prompting.}

\item{backup}{Logical. If TRUE, creates a backup of the right directory before synchronization. The backup is stored in the location specified by \code{backup_dir}.}

\item{backup_dir}{Path to the directory where the backup of the original right directory will be stored. If not specified, the backup is stored in temporary directory (\code{tempdir}).}

\item{overwrite}{Logical. If TRUE (default), existing files at the destination are overwritten. If FALSE, existing destination files are preserved and the copy is skipped.}

\item{verbose}{logical. If TRUE, display directory tree before and after synchronization. Default is FALSE}
}
\value{
Invisible TRUE indicating successful synchronization.
}
\description{
update non common files in right directory based on left one -i.e., the function will:
\itemize{
\item for common_files:
\itemize{
\item do nothing, left unchanged
}
\item for non common files,
\itemize{
\item copy those files that are only in left to right
\item keep in right those files that are only in right (i.e., files 'missing in left')
}
}
}
\examples{
# Create a temporary synchronization environment
\donttest{
e <- toy_dirs()
left  <- e$left
right <- e$right

# Partially update missing files asymmetrically (left → right)
# Option 1: provide left and right paths
partial_update_missing_files_asym_to_right(
  left_path  = left,
  right_path = right
)

# Option 2: provide a precomputed sync_status object
sync_status <- compare_directories(
  left_path  = left,
  right_path = right
)
partial_update_missing_files_asym_to_right(sync_status = sync_status)
}
}
