% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{save_sync_status}
\alias{save_sync_status}
\title{Save sync_status file}
\usage{
save_sync_status(dir_path)
}
\arguments{
\item{dir_path}{path to directory}
}
\value{
the file is saved in a \verb{_syncdr} subdirectory within the specified directory
}
\description{
Save sync_status file
}
\examples{
\donttest{
# Set the directory path
e = toy_dirs()
left <- e$left

save_sync_status(dir_path = left)
}
}
