% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncdr-package.R
\docType{package}
\name{syncdr-package}
\alias{syncdr}
\alias{syncdr-package}
\title{syncdr: Facilitate File Handling, Directory Comparison & Synchronization}
\description{
Compare directories flexibly (by date, content, or both) and synchronize files efficiently, with asymmetric and symmetric modes, helper tools, and visualization support for file management.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rossanatat.github.io/syncdr/}
  \item \url{https://github.com/RossanaTat/syncdr}
  \item Report bugs at \url{https://github.com/RossanaTat/syncdr/issues}
}

}
\author{
\strong{Maintainer}: Rossana Tatulli \email{rtatulli@worldbank.org}

Authors:
\itemize{
  \item R.Andres Castaneda \email{acastanedaa@worldbank.org}
}

Other contributors:
\itemize{
  \item Global Poverty and Inequality Data Team World Bank [copyright holder]
}

}
\keyword{internal}
