% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_sync_status_arg}
\alias{validate_sync_status_arg}
\title{Validate a sync_status argument}
\usage{
validate_sync_status_arg(sync_status)
}
\arguments{
\item{sync_status}{The value to validate.}
}
\value{
Invisibly returns \code{sync_status} if the check passes.
}
\description{
Checks that a \code{sync_status} argument is an object of class \code{"syncdr_status"}.
Throws an informative \code{cli::cli_abort()} if not.
}
\keyword{internal}
