% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymmetric_sync.R
\name{common_files_asym_sync_to_right}
\alias{common_files_asym_sync_to_right}
\title{Partial asymmetric synchronization to right (update common files)}
\usage{
common_files_asym_sync_to_right(
  sync_status,
  by_date = TRUE,
  by_content = FALSE,
  recurse = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{sync_status}{object of class 'syncdr_status' with info on sync status and comparison of directories}

\item{by_date}{logical, TRUE by default}

\item{by_content}{logical, FALSE by default}

\item{recurse}{logical, TRUE by default.
If recurse is TRUE: when copying a file from source folder to destination folder, the file will be copied into the corresponding (sub)directory.
If the sub(directory) where the file is located does not exist in destination folder (or you are not sure), set recurse to FALSE,
and the file will be copied at the top level}
}
\value{
print "synchronized"
}
\description{
Partially synchronize right directory based on left one -i.e., the function will:
\itemize{
\item for common_files:
\itemize{
\item if by date only: copy files that are newer in left to right
\item if by date and content: copy files that are newer and different in left to right
\item if by content only: copy files that are different in left to right
}
\item for non common files, nothing changes: i.e.,
\itemize{
\item disregard those files that are only in left
\item keep in right those files that are only in right (i.e., files 'missing in left')
}
}
}
