% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_directories.R
\name{compare_directories}
\alias{compare_directories}
\title{Compare directories
This function takes two directories as input, say left and right, and compares them.
The goal is to return the status of synchronization -at file level- of the two directories.
The sync status is returned for both:
*1. NON common files, i.e., files that are either only in left or only in right
*2. Common files, i.e., files that are in both directories}
\usage{
compare_directories(
  left_path,
  right_path,
  recurse = TRUE,
  by_date = TRUE,
  by_content = FALSE,
  ...
)
}
\arguments{
\item{left_path}{path of one directory}

\item{right_path}{path of another directory}

\item{recurse}{If TRUE recurse fully, if a positive number the number of levels to recurse.}

\item{by_date}{logical: if TRUE, i.e., the default, it compares the directories based on date of modification of their common files}

\item{by_content}{logical: default is FALSE.
If TRUE, it compares the directories based on whether (hashed) content of common files is same or different (ADD EXPLANATION OF 3 TYPES)}
}
\value{
list of class "syncdr_status", with 4 elements:
\itemize{
\item Non-common files: paths and sync status of files only in right/only in left
\item Common files: paths and sync status of files in both directories
\item Path of left directory
\item Path of right directory
}
}
\description{
Compare directories
This function takes two directories as input, say left and right, and compares them.
The goal is to return the status of synchronization -at file level- of the two directories.
The sync status is returned for both:
*1. NON common files, i.e., files that are either only in left or only in right
*2. Common files, i.e., files that are in both directories
}
\section{possible types of sync status}{


Sync status is computed for both files that are available in both directories
and files that are only into one directory or into the other. It is also computed either by date only,
or by content only, or by date and content -depending on what you choose.

Possible value of sync status -for common files:
\itemize{
\item When comparing by date: new, old, same
\item When comparing by date $ content: | date | content |
new  and diff;
new  and same;
old  and diff;
old  and same;
same and diff;
same and same
\item When comparing by content only: diff or same
}

Possible value of sync status -for non common files:
\itemize{
\item When comparing by date: missing in left or only in right
}
}

