% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{compare_file_contents}
\alias{compare_file_contents}
\title{Compare contents of two files and determine their sync status}
\usage{
compare_file_contents(path_left, path_right)
}
\description{
Compare contents of two files and determine their sync status
}
\keyword{internal}
