% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{compare_modification_times}
\alias{compare_modification_times}
\title{Compare date of last modification of two files and determine sync status}
\usage{
compare_modification_times(modification_time_left, modification_time_right)
}
\description{
Compare date of last modification of two files and determine sync status
}
\keyword{internal}
