% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action_functions.R
\name{copy_files_to_left}
\alias{copy_files_to_left}
\title{Copy files from right directory to left directory}
\usage{
copy_files_to_left(left_dir, right_dir, files_to_copy, recurse = TRUE)
}
\arguments{
\item{left_dir}{path of left directory}

\item{right_dir}{path of right directory}

\item{files_to_copy}{df containing paths of files to copy and their sync status}

\item{recurse}{logical, TRUE by default.
If recurse is TRUE: when copying a file from source folder to destination folder, the file will be copied into the corresponding (sub)directory.
If the sub(directory) where the file is located does not exist in destination folder (or you are not sure), set recurse to FALSE,
and the file will be copied at the top level}
}
\value{
invisible TRUE
}
\description{
This function first checks source and destination path, and build destination path if not available,
then copies the files accordingly
}
\keyword{internal}
