% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{directory_info}
\alias{directory_info}
\title{Get directory information}
\usage{
directory_info(dir, recurse = TRUE, ...)
}
\arguments{
\item{dir}{character string, path of directory}

\item{recurse}{If TRUE recurse fully, if a positive number the number of levels to recurse}
}
\value{
data frame with info on directory's files
}
\description{
Get directory information
}
\keyword{internal}
