% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_comparison.R
\name{display_sync_status}
\alias{display_sync_status}
\title{Display status of synchronization/comparison info between two directories in DT table}
\usage{
display_sync_status(sync_status)
}
\arguments{
\item{sync_status}{object of class \code{"syncdr_status"}, result of \code{compare_directories()}}
}
\value{
DT table showing the comparison between the two directories
together with their synchronization status
}
\description{
Display status of synchronization/comparison info between two directories in DT table
}
\examples{
# Compare directories with 'compare_directories()'
sync_status <- compare_directories(left_path, right_path)
display_sync_status(sync_status$common_files)
display_sync_status(sync_status$non_common_files)

sync_status_date_cont <- compare_directories(left,
                                             right,
                                             by_content = TRUE)
display_sync_status(sync_status_date_cont$common_files)
display_sync_status(sync_status_date_cont$non_common_files)

sync_status_content <- compare_directories(left,
                                           right,
                                           by_content = TRUE,
                                           by_date = FALSE)
display_sync_status(sync_status_content$common_files)
display_sync_status(sync_status_content$non_common_files)


}
