% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_functions.R
\name{display_sync_status}
\alias{display_sync_status}
\title{Display status of synchronization/comparison info between two directories in DT table}
\usage{
display_sync_status(sync_status)
}
\arguments{
\item{sync_status}{object of class \code{"syncdr_status"}, result of \code{compare_directories()}}
}
\value{
DT table showing the comparison between the two directories
together with their synchronization status
}
\description{
Display status of synchronization/comparison info between two directories in DT table
}
