% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{filter_common_files}
\alias{filter_common_files}
\title{Filter files that are common between the two directories under comparison}
\usage{
filter_common_files(
  sync_status,
  by_date = TRUE,
  by_content = FALSE,
  dir = "left"
)
}
\arguments{
\item{sync_status}{object of class 'syncdr_status' with info on sync status
and comparison of directories}

\item{by_date}{logical, TRUE by default}

\item{by_content}{logical, FALSE by default}
}
\value{
'syncdr_status' object filtered accordingly
}
\description{
The filtering is done by date, by date and content or by content only
}
\keyword{internal}
