% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{filter_non_common_files}
\alias{filter_non_common_files}
\title{Filter files that are NOT common between the two directories under comparison}
\usage{
filter_non_common_files(sync_status, dir = "left")
}
\arguments{
\item{sync_status}{object of class 'syncdr_status' with info on sync status
and comparison of directories}

\item{dir}{character, either "left", "right", "all" (both directories)}
}
\value{
'syncdr_status' object filtered accordingly
}
\description{
This function filters non common files in "syncdr_status" object (resulting from 'compare_directories()')
}
\keyword{internal}
