% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymmetric_sync.R
\name{full_asym_sync_to_right}
\alias{full_asym_sync_to_right}
\title{Full asymmetric synchronization to right}
\usage{
full_asym_sync_to_right(sync_status, by_date = TRUE, by_content = FALSE)
}
\arguments{
\item{sync_status}{object of class 'syncdr_status' with info on sync status and comparison of directories}

\item{by_date}{logical, TRUE by default}

\item{by_content}{logical, FALSE by default}
}
\value{
print "synchronized"
}
\description{
Fully synchronize right directory based on left one -i.e., the function will:
\itemize{
\item for common_files:
\itemize{
\item if by date only: copy files that are newer in left to right
\item if by date and content: copy files that are newer and different in left to right
\item if by content only: copy files that are different in left to right
}
\item copy to right those files that are only in left
\item delete in right those files that are only in right (i.e., files 'missing in left')
}
}
