% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymmetric_sync.R
\name{full_asym_sync_to_right}
\alias{full_asym_sync_to_right}
\title{Full asymmetric synchronization to right directory}
\usage{
full_asym_sync_to_right(
  left_path,
  right_path,
  by_date = TRUE,
  by_content = FALSE,
  recurse = TRUE,
  verbose = getOption("syncdr.verbose")
)
}
\arguments{
\item{left_path}{Path to the left/first directory.}

\item{right_path}{Path to the right/second directory.}

\item{by_date}{Logical. If TRUE, synchronize based on file modification dates (default is TRUE).}

\item{by_content}{Logical. If TRUE, synchronize based on file contents (default is FALSE).}

\item{recurse}{Logical. If TRUE (default), files are copied to corresponding subdirectories
in the destination folder. If FALSE, files are copied to the top level of the destination folder
without creating subdirectories if they do not exist.}

\item{verbose}{logical. If TRUE, display directory tree before and after synchronization. Default is FALSE}
}
\value{
Invisible TRUE indicating successful synchronization.
}
\description{
This function performs a full asymmetric synchronization of the right directory
based on the left directory. It includes the following synchronization steps:
}
\details{
\itemize{
\item For common files:
\itemize{
\item If comparing by date only (\code{by_date = TRUE}): Copy files that are newer in the left directory to the right directory.
\item If comparing by date and content (\code{by_date = TRUE} and \code{by_content = TRUE}): Copy files that are newer and different in the left directory to the right directory.
\item If comparing by content only (\code{by_content = TRUE}): Copy files that are different in the left directory to the right directory.
}
\item Copy to the right directory those files that exist only in the left directory.
\item Delete from the right directory those files that are exclusive in the right directory (i.e., missing in the left directory)
}
}
\examples{
# Compare directories with 'compare_directories()'
e <- toy_dirs()

# Get left and right directories' paths
left  <- e$left
right <- e$right
# Synchronize by date & content
full_asym_sync_to_right(left_path  = left,
                        right_path = right,
                        by_content = TRUE)
}
