% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric_sync.R
\name{full_symmetric_sync}
\alias{full_symmetric_sync}
\title{Full symmetric synchronization}
\usage{
full_symmetric_sync(
  sync_status,
  by_date = TRUE,
  by_content = FALSE,
  recurse = TRUE,
  verbose = getOption("syncdr.verbose")
)
}
\arguments{
\item{sync_status}{object of class 'syncdr_status' with info on sync status and comparison of directories}

\item{by_date}{logical, TRUE by default}

\item{by_content}{logical, FALSE by default}

\item{recurse}{logical, TRUE by default.
If recurse is TRUE: when copying a file from source folder to destination folder, the file will be copied into the corresponding (sub)directory.
If the sub(directory) where the file is located does not exist in destination folder (or you are not sure), set recurse to FALSE,
and the file will be copied at the top level}

\item{verbose}{logical. If TRUE, display directory tree before and after synchronization. Default is FALSE}
}
\value{
print "synchronized"
}
\description{
This function updates directories in the following way:
\itemize{
\item For common files:
\itemize{
\item if by date: If the file in one directory is newer than the corresponding file in the other directory,
it will be copied over to update the older version. If modification dates are the same, no action is taken
\item if by date and content: If the file in one directory is newer AND different than the corresponding file in the other directory,
it will be copied over to update the older version. If modification dates/contents are the same, no action is taken
\item if by content only: ? TO DECIDE WHAT TO DO WITH THOSE FILES
}
\item For non common files:
\itemize{
\item if a file exists in one but not in the other it is copied to the other directory
}
}
}
