% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{hash_files}
\alias{hash_files}
\title{Hash files by content}
\usage{
hash_files(files_path)
}
\arguments{
\item{files_path}{character vector of paths of files to hash}
}
\value{
hashes of files
}
\description{
Hash files by content
}
\keyword{internal}
