% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.syncdr_status}
\alias{print.syncdr_status}
\title{Title}
\usage{
\method{print}{syncdr_status}(x, ...)
}
\arguments{
\item{x}{obbject of syncdr_status class created in \link{compare_directories}}

\item{...}{}
}
\value{
prints syncdr_status object
}
\description{
Title
}
