% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{save_sync_status}
\alias{save_sync_status}
\title{Save sync_status file}
\usage{
save_sync_status(dir_path)
}
\arguments{
\item{dir_path}{path to directory}
}
\value{
a file storing a summary of the sync_status, saved in XXXXX TBC
}
\description{
Save sync_status file
}
