% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{search_duplicates}
\alias{search_duplicates}
\title{Search for duplicate files
-by date& content or by content only}
\usage{
search_duplicates(dir_path, verbose = TRUE)
}
\description{
Search for duplicate files
-by date& content or by content only
}
