% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_aux_fun.R
\name{style_msgs}
\alias{style_msgs}
\title{Apply custom style to text}
\usage{
style_msgs(color_name, text)
}
\arguments{
\item{color_name}{name of color to apply (only selected choices for now)}
}
\value{
printed text
}
\description{
Apply custom style to text
}
\keyword{internal}
