% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncdr-package.R
\docType{package}
\name{syncdr-package}
\alias{syncdr}
\alias{syncdr-package}
\title{syncdr: Tool for Facilitating Directory Comparison and Updating}
\description{
Compare directories flexibly (by date, content, or both) and synchronize files efficiently, with asymmetric and symmetric modes, helper tools, and visualization support for file management.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rossanatat.github.io/syncdr/}
}

}
\author{
\strong{Maintainer}: R.Andres Castaneda \email{acastanedaa@worldbank.org}

Authors:
\itemize{
  \item Rossana Tatulli \email{rtatulli@worldbank.org}
}

}
\keyword{internal}
