% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncdr-package.R
\docType{package}
\name{syncdr-package}
\alias{syncdr}
\alias{syncdr-package}
\title{syncdr: Tool for Facilitating Directory Comparison and Updating}
\description{
An R package for directory comparison and updates. It provides tools to facilitate file handling, comparison and synchronization of directories.
}
\author{
\strong{Maintainer}: R.Andres Castaneda \email{acastanedaa@worldbank.org}

Authors:
\itemize{
  \item Rossana Tatulli \email{rtatulli@worldbank.org}
}

}
\keyword{internal}
