% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_dirs.R
\name{toy_dirs}
\alias{toy_dirs}
\title{Create tow directories to test syncdr functions}
\usage{
toy_dirs(verbose = FALSE)
}
\arguments{
\item{verbose}{logical: display information. Default is FALSE}
}
\value{
list with directory paths for left and right
}
\description{
Create tow directories to test syncdr functions
}
\examples{
l <- toy_dirs()
compare_directories(l$left, l$right)
}
