% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymmetric_sync.R
\name{update_missing_files_asym_to_right}
\alias{update_missing_files_asym_to_right}
\title{Full asymmetric asymmetric synchronization of non common files}
\usage{
update_missing_files_asym_to_right(
  sync_status,
  by_date = TRUE,
  by_content = FALSE
)
}
\arguments{
\item{sync_status}{object of class 'syncdr_status' with info on sync status and comparison of directories}

\item{by_date}{logical, TRUE by default}

\item{by_content}{logical, FALSE by default}
}
\value{
print "synchronized"
}
\description{
update non common files in right directory based on left one -i.e., the function will:
\itemize{
\item for common_files:
\itemize{
\item do nothing
}
\item for non common files,
\itemize{
\item copy those files that are only in left to right
\item delete in right those files that are only in right (i.e., files 'missing in left')
}
}
}
